fix: fix js
