+++ packages/functions-runtime/src/parsing.js
@@ -1,5 +1,6 @@
 const { Duration } = require("./Duration");
 const { InlineFile, File } = require("./File");
+const { isPlainObject } = require("./type-utils");
 
 // parseInputs takes a set of inputs and creates objects for the ones that are of a complex type.
 //
@@ -81,18 +82,6 @@ function transformRichDataTypes(data) {
   return row;
 }
 
-function isPlainObject(obj) {
-  return Object.prototype.toString.call(obj) === "[object Object]";
-}
-
-function isRichType(obj) {
-  if (!isPlainObject(obj)) {
-    return false;
-  }
-
-  return obj instanceof Duration;
-}
-
 function isReferencingExistingRecord(value) {
   return Object.keys(value).length === 1 && value.id;
 }
@@ -101,7 +90,5 @@ module.exports = {
   parseInputs,
   parseOutputs,
   transformRichDataTypes,
-  isPlainObject,
-  isRichType,
   isReferencingExistingRecord,
 };
